fix : 파일 이동및 파일 분할
